refactor(covenantsigner): split monolithic validation.go into focused domain files
